feat: fix tests
