+++ apps/namadillo/src/atoms/balance/functions.ts
@@ -1,67 +1,10 @@
-import { Asset } from "@chain-registry/types";
 import { IbcToken, NativeToken } from "@namada/indexer-client";
 import { mapCoinsToAssets } from "atoms/integrations";
 import BigNumber from "bignumber.js";
-import * as osmosis from "chain-registry/mainnet/osmosis";
 import { DenomTrace } from "cosmjs-types/ibc/applications/transfer/v1/transfer";
-import { Address, AddressWithAssetAndAmountMap } from "types";
+import { AddressWithAssetAndAmountMap } from "types";
 import { isNamadaAsset } from "utils";
 import { TokenBalance } from "./atoms";
-import { fetchCoinPrices } from "./services";
-
-// TODO upgrade this function to be as smart as possible
-// please refer to `tryCoinToIbcAsset` and how we could combine both
-export const findAssetByToken = (
-  token: NativeToken | IbcToken,
-  assets: Asset[]
-): Asset | undefined => {
-  if ("trace" in token) {
-    const traceDenom = token.trace.split("/").at(-1);
-    if (traceDenom) {
-      for (let i = 0; i < assets.length; i++) {
-        const asset = assets[i];
-        if (
-          asset.base === traceDenom ||
-          asset.denom_units.find(
-            (u) => u.denom === traceDenom || u.aliases?.includes(traceDenom)
-          )
-        ) {
-          return asset;
-        }
-      }
-    }
-  }
-  return undefined;
-};
-
-export const fetchTokenPrices = async (
-  tokenAddressToFetch: Address[],
-  apiTokens: (NativeToken | IbcToken)[]
-): Promise<Record<Address, BigNumber>> => {
-  const baseMap: Record<string, string> = {};
-  tokenAddressToFetch.forEach((address) => {
-    const token = apiTokens?.find((t) => t.address === address);
-    if (token) {
-      // searching only on osmosis because these are the assets supported by fetchCoinPrices
-      const asset = findAssetByToken(token, osmosis.assets.assets);
-      if (asset) {
-        baseMap[asset.base] = address;
-      }
-    }
-  });
-  const baseList = Object.keys(baseMap);
-  const apiResponse = await fetchCoinPrices(baseList);
-
-  const tokenPrices: Record<string, BigNumber> = {};
-  Object.entries(apiResponse).forEach(([base, value]) => {
-    const address = baseMap[base];
-    const dollar = Object.values(value)[0];
-    if (dollar) {
-      tokenPrices[address] = new BigNumber(dollar);
-    }
-  });
-  return tokenPrices;
-};
 
 /**
  * Sum the dollar amount of a list of tokens
@@ -88,9 +31,9 @@ export const getTotalNam = (list?: TokenBalance[]): BigNumber =>
 
 const tnamAddressToDenomTrace = (
   address: string,
-  tokenAddresses: (NativeToken | IbcToken)[]
+  chainTokens: (NativeToken | IbcToken)[]
 ): DenomTrace | undefined => {
-  const token = tokenAddresses.find((entry) => entry.address === address);
+  const token = chainTokens.find((entry) => entry.address === address);
   const trace = token && "trace" in token ? token.trace : undefined;
 
   // If no trace, the token is NAM, but return undefined because we only care
@@ -113,7 +56,7 @@ const tnamAddressToDenomTrace = (
 
 export const mapNamadaAddressesToAssets = async (
   balances: { address: string; minDenomAmount: BigNumber }[],
-  tokenAddresses: (NativeToken | IbcToken)[],
+  chainTokens: (NativeToken | IbcToken)[],
   chainId: string
 ): Promise<AddressWithAssetAndAmountMap> => {
   const coins = balances.map(({ address, minDenomAmount }) => ({
@@ -122,7 +65,7 @@ export const mapNamadaAddressesToAssets = async (
   }));
 
   return await mapCoinsToAssets(coins, chainId, (tnamAddress) =>
-    Promise.resolve(tnamAddressToDenomTrace(tnamAddress, tokenAddresses))
+    Promise.resolve(tnamAddressToDenomTrace(tnamAddress, chainTokens))
   );
 };
