+++ apps/namadillo/src/atoms/integrations/functions.ts
@@ -2,7 +2,7 @@ import { Asset, AssetList, Chain, IBCInfo } from "@chain-registry/types";
 import { QueryClient, setupIbcExtension } from "@cosmjs/stargate";
 import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { Account, IbcTransferProps } from "@namada/types";
-import { mapUndefined, shortenAddress } from "@namada/utils";
+import { mapUndefined } from "@namada/utils";
 import BigNumber from "bignumber.js";
 import * as celestia from "chain-registry/mainnet/celestia";
 import * as cosmos from "chain-registry/mainnet/cosmoshub";
@@ -28,6 +28,7 @@ import {
   LocalnetToml,
 } from "types";
 import { toBaseAmount, toDisplayAmount } from "utils";
+import { unknownAsset } from "utils/assets";
 import { getSdkInstance } from "utils/sdk";
 
 import housefireAssets from "namada-chain-registry/_testnets/namadahousefire/assetlist.json";
@@ -212,19 +213,6 @@ const tryDenomToIbcAsset = async (
 
   return originalChainRegistryAsset || unknownAsset(path + "/" + baseDenom);
 };
-
-export const unknownAsset = (denom: string): Asset => ({
-  denom_units: [
-    {
-      denom,
-      exponent: 0,
-    },
-  ],
-  base: denom,
-  name: denom,
-  display: denom,
-  symbol: shortenAddress(denom, 4, 4),
-});
 
 const findOriginalAsset = async (
   coin: Coin,
